Subscriptions, reload and restore buttons, fixes
Submissions are now stored in localStorage, code is encrypted.
Code editor is now running properly, hotkeys should work correctly, minimap disabled.
Student can restore last evaluation.
It is possible now to change the terminal font size and zoom in/out the entire app (this feature still needs to be adjusted).
Added function to clear localStorage of data older than 30 days.
If Student has no rewards the app should display a proper message. Rewards of "UNLOCK" and "REVEAL" kind are not displayed.
Added "Waiting for result" state to the terminal and icon that indicates current state of the connection. If there are any issues with the connection or after 120 seconds of waiting and no response app will display a red icon.
Initial code for exercises is added if exists.